Save ride name and mapId in database

+++ server/db/tables.sql
@@ -30,8 +30,8 @@ CREATE TABLE points (
 CREATE TABLE rides (
   id SERIAL PRIMARY KEY NOT NULL,
   name VARCHAR(255),
-  distance smallint,
-  speed smallint,
+  distance INTEGER,
+  speed INTEGER,
   isSelected boolean DEFAULT false,
   isActive boolean DEFAULT true,
   createdBy INTEGER REFERENCES users(id) ON DELETE CASCADE,
